+++ changelog.md
@@ -14,6 +14,7 @@ latest
  - 'Retain session data between browser refreshes?' setting on settings panel.
  - Gain adjustment on the edit panel (Thanks to eljeff).
  - Option in settings panel to download single files when Shift+Clicked to prevent accidental downloads being triggered.
+ - Added rough stretch to selected actions list, this attempts to retain pitch while doubling the duration of the sample.
 
 2023.10.15 1.4.8
  - New serialize selected action, any selected samples that are stereo files, will have their stereo channels serialized to mono, by Left/Right, Side/Mid, or Left/Right/Side/Mid.